+++ tests/test_telephony_call_recording.py
@@ -9,7 +9,6 @@ from wxc_sdk.telephony.call_recording import CallRecordingRegion, LocationCompli
 
 
 class TestCallRecording(TestWithLocations):
-
     def test_read_recording_info(self):
         info = self.api.telephony.call_recording.read()
         print(json.dumps(info.model_dump(mode='json'), indent=2))
@@ -23,14 +22,26 @@ class TestCallRecording(TestWithLocations):
         ca = self.api.telephony.call_recording.read_org_compliance_announcement()
         print(json.dumps(ca.model_dump(mode='json'), indent=2))
 
+    def test_read_org_call_recording_announcement(self):
+        ca = self.api.telephony.call_recording.read_org_call_recording_announcement()
+        print(json.dumps(ca.model_dump(mode='json'), indent=2))
+
     @async_test
     async def test_read_location_compliance_announcement(self):
+        """
+        Retrieve location compliance announcement settings for all calling locations
+        """
         results = await asyncio.gather(
-            *[self.async_api.telephony.call_recording.read_location_compliance_announcement(location_id=loc.location_id)
-              for loc in self.locations],
-            return_exceptions=True)
+            *[
+                self.async_api.telephony.call_recording.read_location_compliance_announcement(
+                    location_id=loc.location_id
+                )
+                for loc in self.locations
+            ],
+            return_exceptions=True,
+        )
         err = None
-        for location, result in zip(self.locations, results):
+        for location, result in zip(self.locations, results):  # noqa: B905
             location: Location
             if isinstance(result, Exception):
                 err = err or result
@@ -46,22 +57,36 @@ class TestCallRecording(TestWithLocations):
 
 class TestSettings(TestCaseWithLog):
     def test_list_regions(self):
+        """
+        List call recording regions
+        """
         regions = self.api.telephony.call_recording.get_call_recording_regions()
         print(f'got {len(regions)} regions')
         print(json.dumps(TypeAdapter(list[CallRecordingRegion]).dump_python(regions, mode='json'), indent=2))
 
     def test_get_org_vendors(self):
+        """
+        get call recording vendors
+        :return:
+        """
         vendors = self.api.telephony.call_recording.get_org_vendors()
         print(json.dumps(vendors.model_dump(mode='json', by_alias=True), indent=2))
 
     def test_get_org_vendor_id(self):
+        """
+        Get org vendor id
+        """
         vendors = self.api.telephony.call_recording.get_org_vendors()
         print(f'got vendor id: {vendors.vendor_id}')
+        print(f'   vendor name: {vendors.vendor_name}')
 
     def test_org_users(self):
+        """
+        Get call recording users
+        :return:
+        """
         users = list(self.api.telephony.call_recording.list_org_users())
         print(f'got {len(users)} users')
 
     @async_test
-    async def location_vendor(self):
-        ...
+    async def location_vendor(self): ...