+++ wxc_sdk/telephony/call_recording/__init__.py
@@ -1,12 +1,13 @@
 from collections.abc import Generator
-from typing import Optional
+from typing import Any, Optional
 
 from pydantic import Field, TypeAdapter
 
 from wxc_sdk.api_child import ApiChild
 from wxc_sdk.base import ApiModel, enum_str
 from wxc_sdk.base import SafeEnum as Enum
-from wxc_sdk.common import IdAndName, UserLicenseType, UserType
+from wxc_sdk.common import AnnAudioFile, IdAndName, UserLicenseType, UserType
+from wxc_sdk.person_settings.call_recording import BaseCallRecordingAnnouncement, CallRecordingAnnouncement
 
 __all__ = [
     'CallRecordingInfo',
@@ -51,7 +52,7 @@ class CallRecordingTermsOfService(ApiModel):
     terms_of_service_url: Optional[str] = None
 
 
-class OrgComplianceAnnouncement(ApiModel):
+class BaseComplianceAnnouncement(ApiModel):
     #: Flag to indicate whether the Call Recording START/STOP announcement is played to an inbound caller.
     inbound_pstncalls_enabled: Optional[bool] = Field(alias='inboundPSTNCallsEnabled', default=None)
     #: Flag to indicate whether the Call Recording START/STOP announcement is played to an outbound caller.
@@ -62,9 +63,50 @@ class OrgComplianceAnnouncement(ApiModel):
     delay_in_seconds: Optional[int] = None
 
 
-class LocationComplianceAnnouncement(OrgComplianceAnnouncement):
+class OrgComplianceAnnouncement(BaseComplianceAnnouncement):
+    #: Flag to indicate whether to use the custom compliance announcement. If true it uses the organization's custom
+    #: compliance announcement file, and if false default compliance announcement used.
+    use_custom_announcement_enabled: Optional[bool] = None
+    #: The custom audio announcement file to be played.
+    audio_announcement_file: Optional[AnnAudioFile] = None
+
+    def update(self) -> dict[str, Any]:
+        data = self.model_dump(mode='json', by_alias=True, exclude_none=True, exclude={'audio_announcement_file'})
+        if self.use_custom_announcement_enabled:
+            if (af := self.audio_announcement_file) is None or af.id is None:
+                raise ValueError('audio announcement file id is required')
+            data['audioAnnouncementFileId'] = af.id
+        return data
+
+
+class LocationComplianceAnnouncement(BaseComplianceAnnouncement):
     #: Flag to indicate whether to use the customer level compliance announcement default settings.
     use_org_settings_enabled: Optional[bool] = None
+    #: Flag to indicate whether to use the organization level custom compliance announcement. If this flag is set to
+    #: true, takes the organization's announcement setting. If this flag is set to false, takes the location's custom
+    #: announcement.
+    use_org_level_announcement_enabled: Optional[bool] = None
+    #: Custom compliance announcement settings.
+    custom_compliance_announcement: Optional[CallRecordingAnnouncement] = None
+
+    def update(self) -> dict[str, Any]:
+        """
+        Data for update
+
+        :meta private:
+        """
+        data = self.model_dump(mode='json', by_alias=True, exclude={'custom_compliance_announcement'})
+        if (ola := self.use_org_level_announcement_enabled) is not None and not ola:
+            if (
+                (cca := self.custom_compliance_announcement) is None
+                or (af := cca.audio_announcement_file) is None
+                or af.id is None
+            ):
+                raise ValueError(
+                    'custom compliance announcement is required and audio announcement file id is required'
+                )
+            data['customComplianceAnnouncement'] = {'type': cca.type, 'audioAnnouncementFileId': af.id}
+        return data
 
 
 class CallRecordingRegion(ApiModel):
@@ -278,10 +320,60 @@ class CallRecordingSettingsApi(ApiChild, base='telephony/config'):
         url = self.ep(f'callRecording/vendors/{vendor_id}/termsOfService')
         super().put(url, params=params, json=body)
 
+    def read_org_call_recording_announcement(self, org_id: str = None) -> BaseCallRecordingAnnouncement:
+        """
+        Get Organization Call Recording Announcement Settings
+
+        Retrieve the organization call recording announcements setting.
+
+        The call recording Announcement feature interacts with the Call Recording feature, specifically with the
+        playback of the start/stop announcement. When the compliance announcement is played to the PSTN party, and the
+        PSTN party is connected to a party with call recording enabled, then the start/stop announcement is inhibited.
+
+        Retrieving organization compliance announcement setting requires a full or read-only administrator auth token
+        with a scope of `spark-admin:telephony_config_read`.
+
+        :param org_id: Retrieve call recording announcements setting from this organization.
+        :type org_id: str
+        :rtype: :class:`CallRecordingAnnouncements`
+        """
+        params: dict[str, Any] = dict()
+        if org_id is not None:
+            params['orgId'] = org_id
+        url = self.ep('callRecording/announcements')
+        data = super().get(url, params=params)
+        r = BaseCallRecordingAnnouncement.model_validate(data)
+        return r
+
+    def update_org_call_recording_announcement(
+        self, settings: BaseCallRecordingAnnouncement, org_id: str = None
+    ) -> None:
+        """
+        Update Organization Call Recording Announcement Settings.
+
+        The Compliance Announcement feature interacts with the Call Recording feature, specifically with the playback
+        of the start/stop announcement. When the compliance announcement is played to the PSTN party, and the PSTN
+        party is connected to a party with call recording enabled, then the start/stop announcement is inhibited.
+
+        Updating the organization compliance announcement requires a full administrator auth token with a scope of
+        `spark-admin:telephony_config_write`.
+
+        :param settings: Update compliance announcement settings for this organization.
+        :type settings: BaseCallRecordingAnnouncement
+        :param org_id: Update the call recording announcements setting from this organization.
+        :type org_id: str
+        :rtype: None
+        """
+        params: dict[str, Any] = dict()
+        if org_id is not None:
+            params['orgId'] = org_id
+        body = settings.update()
+        url = self.ep('callRecording/announcements')
+        super().put(url, params=params, json=body)
+
     def read_org_compliance_announcement(self, org_id: str = None) -> OrgComplianceAnnouncement:
         """
         Get details for the organization Compliance Announcement Setting
-
 
         Retrieve the organization compliance announcement settings.
 
@@ -304,9 +396,13 @@ class CallRecordingSettingsApi(ApiChild, base='telephony/config'):
         r = OrgComplianceAnnouncement.model_validate(data)
         return r
 
-    def update_org_compliance_announcement(self, settings: OrgComplianceAnnouncement, org_id: str = None):
+    def update_org_compliance_announcement(
+        self,
+        settings: OrgComplianceAnnouncement,
+        org_id: str = None,
+    ):
         """
-        Update the organization compliance announcement
+        Update the organization compliance announcement.
 
         The Compliance Announcement feature interacts with the Call Recording feature, specifically with the playback
         of the start/stop announcement. When the compliance announcement is played to the PSTN party, and the PSTN
@@ -324,7 +420,7 @@ class CallRecordingSettingsApi(ApiChild, base='telephony/config'):
         params = {}
         if org_id is not None:
             params['orgId'] = org_id
-        body = settings.model_dump(mode='json', by_alias=True, exclude_none=True)
+        body = settings.update()
         url = self.ep('callRecording/complianceAnnouncement')
         super().put(url, params=params, json=body)
 
@@ -358,10 +454,13 @@ class CallRecordingSettingsApi(ApiChild, base='telephony/config'):
         return r
 
     def update_location_compliance_announcement(
-        self, location_id: str, settings: LocationComplianceAnnouncement, org_id: str = None
-    ):
+        self,
+        location_id: str,
+        settings: LocationComplianceAnnouncement,
+        org_id: str = None,
+    ) -> None:
         """
-        Update the location compliance announcement
+        Update the location compliance announcement.
 
         The Compliance Announcement feature interacts with the Call Recording feature, specifically with the playback
         of the start/stop announcement. When the compliance announcement is played to the PSTN party, and the PSTN
@@ -381,7 +480,7 @@ class CallRecordingSettingsApi(ApiChild, base='telephony/config'):
         params = {}
         if org_id is not None:
             params['orgId'] = org_id
-        body = settings.model_dump(mode='json', by_alias=True, exclude_none=True)
+        body = settings.update()
         url = self.ep(f'locations/{location_id}/callRecording/complianceAnnouncement')
         super().put(url, params=params, json=body)
 
@@ -451,7 +550,7 @@ class CallRecordingSettingsApi(ApiChild, base='telephony/config'):
         Set Call Recording Vendor for a Location
 
         Assign a call recording vendor to a location of an organization. Response will be `204` if the changes can be
-        applied immediatley otherwise `200` with a job ID is returned.
+        applied immediately otherwise `200` with a job ID is returned.
 
         The Call Recording feature supports multiple third-party call recording providers, or vendors, to capture and
         manage call recordings. An organization is configured with an overall provider, but locations can be
@@ -480,7 +579,7 @@ class CallRecordingSettingsApi(ApiChild, base='telephony/config'):
         params = {}
         if org_id is not None:
             params['orgId'] = org_id
-        body = dict()
+        body: dict[str, Any] = dict()
         if id is not None:
             body['id'] = id
         if org_default_enabled is not None:
@@ -525,7 +624,13 @@ class CallRecordingSettingsApi(ApiChild, base='telephony/config'):
         return r
 
     def list_location_users(
-        self, location_id: str, standard_user_only: bool = None, org_id: str = None, **params
+        self,
+        location_id: str,
+        max_: int = None,
+        start: int = None,
+        standard_user_only: bool = None,
+        org_id: str = None,
+        **params,
     ) -> Generator[RecordingUser, None, None]:
         """
         Get Call Recording Vendor Users for a Location
@@ -541,6 +646,10 @@ class CallRecordingSettingsApi(ApiChild, base='telephony/config'):
 
         :param location_id: Retrieve vendor users for this location.
         :type location_id: str
+        :param max_: Limit the number of vendor users returned to this maximum count. The default is 2000.
+        :type max_: int
+        :param start: Start at the zero-based offset in the list of matching objects. The default is 0.
+        :type start: int
         :param standard_user_only: If true, results only include Webex Calling standard users.
         :type standard_user_only: bool
         :param org_id: Retrieve vendor users for this organization.
@@ -548,6 +657,10 @@ class CallRecordingSettingsApi(ApiChild, base='telephony/config'):
         """
         if org_id is not None:
             params['orgId'] = org_id
+        if max_ is not None:
+            params['max'] = max_
+        if start is not None:
+            params['start'] = start
         if standard_user_only is not None:
             params['standardUserOnly'] = str(standard_user_only).lower()
         url = self.ep(f'locations/{location_id}/callRecording/vendorUsers')