+++ docs/user/changes.rst
@@ -2,6 +2,7 @@ Release history
 ===============
 
 - feat: new api api.telephony.callqueue.dnis (call queue DNIS settings)
+- feat: read/update org level call recording announcement settings, api.telephony.call_recording.read_org_call_recording_announcement/update_org_call_recording_announcement
 
 1.36.0
 ------